fix docs regarding TGI endpoint url (#408)
The README specify to use a `/generate_stream` url, when using a custom
TGI endpoint. This is incorrect, as it breaks APIs, such as the
conversation summary that depends on the `/` TGI interface.

This turns out to be a documentation-only bug, as `/` will point to
`/generate_stream` when `stream: true` is set in the request JSON.

Simply using the `/` url as the TGI endpoint therefore fixes the
summarization bug.

Fixes: #278